requirements.txt
bcrypt==4.1.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.1
SQLAlchemy==2.0.23
WTForms==3.1.1
mysql-connector-python==8.3.0
